WIP: no trip state
Kind of lost track of where I was with this.  I removed the current trip
state from the trip service. I may want to do the same with the view
model.

Seems to be mostly working, but it seems like there is a race condition
that prevents the trip table from being updated at the end of the trip.
Can't confirm though.