Sync merges with Gfit
Didn't test this thoroughly but it seems to work in the basic case.

Seems like there might be a Gfit bug. It's hard to tell. For some reason
the merged ride is taking on the wrong name in Gfit. I tried deleting
the trip and reinserting it, but that didn't matter. Maybe it's an
eventual consistency thing.