Replace naive rate limit code
Rate limits are per app, not per client so the request counter does
little good. Handle 429 responses and retry failed requests. This is
still not sufficient as the backoff doesn't know what the usage actually
is. Should store the a value in shared preferences that is the next
valid request time.